requirements.txt
tensorflow-cpu
numpy
streamlit
opencv-contrib-python-headless
streamlit-webrtc
